Database transaction closing
Fix an embarrassing mistake in the way we deal with the database - we were
opening a transaction, then sitting and waiting for user input.  After
user input we would make changes, commit, and then open the next
transaction.

The now() function in SQL returns the time/date at the start of the
transaction...  this could be several minutes earlier!