Configurable hotkeys; exit/restart menu option
Menu options for exit/restart machine/reboot machine, which make the
till software return a different exit code to be picked up by
whichever script starts it.

Add a "hotkeys" parameter to the register page to enable additional
key bindings to be specified in configuration files.  When this
parameter is used, configuration files will have to specify bindings
for "Manage Till", "Manage Stock", "Record Waste", "Price Check" and
"Use Stock" as well, because the defaults will be overridden.