Minor release script improvement
Use the version number in setup.py in the working tree, not the index.
This means that test packages can be generated without having to add
a changed setup.py to the index.  The setup.py from the working tree
replaces the one checked out from the index to make sure the generated
package has the correct version number.