Rearrange 'Use Stock' menu to put most likely stocktypes first
Create a table in the database to store (stocklineid,stocktype) pairs.
Add rules such that whenever an entry is made in the stockonsale table
the (stocklineid,stocktype) pair is recorded.
When picking stock for a stockline, sort the list so that items that
have been sold on that line before go at the top of the list.