Prevent duplicate entries in stock_search
td.stock_search() has an option to search for stock of a type
that has been on sale on a particular stockline in the past;
it bumps these items of stock to the head of the list.  This
was buggy; if a stocktype had been on sale on multiple
different stocklines, it would output extra results.  This
bug is now fixed!