Configuration of department restrictions moved to global config file
Two types of restriction have been modified:
1) 'Modifier' keys (Half, Double, Jug, etc.) were previously hard-coded,
along with a list of departments that were valid for each key.  This
list has been moved to a new configuration file entry, 'modkeyinfo'

2) Restrictions on the prices that could be charged using a department
key used to have to be hard-coded in register.py.  There is now a
function in the configuration file that is called for every
department keypress, 'deptkeycheck', that is passed department and
price and must return None (for 'ok') or a string or list of strings
to be displayed to the user.