Register page can now display multi-line transaction lines
The display code in the register page has been altered to be able to
display transaction lines that take more than one line on the screen.
Word-wrapping is done as appropriate.

There is still some re-factoring and simplification to be done; the
redraw code is called too many times in some places, and bugs remain
with respect to updating the "status line / buffer" section of the
screen.