Remove arbitrary minimum screen width
The code that draws the screen header had a hard-coded minimum screen width
of 50 characters.  That limit has been removed.  If there is too much
information to go in the header, it will be truncated.

Several of the popup windows still have hard-coded minimum sizes.  This
will be much harder to change.