Fix part-payment when keyguard is active
When a transaction has been recalled, the "keyguard" feature adds an extra
popup when the Cash key is pressed, to avoid one of the most common user
errors (pressing 'Cash' after entering drinks on a tab, instead of
hiding the tab away with 'Recall Transaction' followed by 'Cash').
Unfortunately, this was throwing away information about part-payments
too soon.  Re-order the code in register.py/cashkey() so that part-payments
are still possible when the keyguard is active.