New "scrollable" field type; update linepopup and menu to use it
Until now, the general concept of "scrolling" through a list of items
has been implemented in an ad-hoc basis in several places in the code.
This commit is a first move to centralising the code for a "scrollable"
field.  Further commits should remove duplicate scrolling code from
the register, food entry system, and so on.

There are two motivations for this commit:
1) Reducing code duplication
2) A desire to add more UI that involves scrollable lists, with more
   flexibility than that afforded by the existing ui.menu code