New 'tableline' for scrollable; split ui.menu into two classes
Scrollable tables have so far been made using the 'table' class in
ui.py; this is used to generate a list of strings.  Incorporate this
functionality into a new subclass of emptyline() suitable for use with
scrollables.

The menu popup has implicit processing of the 'Enter' key, which is not
always used.  It has now been split into two classes; 'menu' works the
same way as before.  'listpopup' is a new class without the implicit
'Enter' key processing.