Change stock auto-allocate function to use stockline_stocktype_log
The old auto-allocate function used the types of stock currently
allocated to a stockline to select new stock items to add to that line.
This worked ok most of the time, but if we were ever completely out of
stock of a line the association was lost.  We now have the
stockline_stocktype_log table, maintained automatically by the database,
which keeps these associations permanently.

The new auto-allocate function is now efficient enough to be invoked
automatically at delivery confirmation time without a nasty pause in the
UI.  We do so, for just those stock items in the delivery.