Add option to purge finished stock explicitly
Stock items assigned to lines with a display capacity would hang around
forever if they were not regularly purged when empty.  Until now this
has only happened at the close of every session.  There is now an option
to do this on the stock maintenance menu - useful to force a purge
before deleting a stock line that is no longer in use.