Fix up date when session is started just before midnight
It is fairly common for the day's session to be finished shortly after closing time at 11pm,
and then for any sales to staff to be on the next day's session.  If the new session is
started between 11pm and midnight, set the default date to be tomorrow.  This is going to be
correct 99% of the time!