Share transaction line to string code between register and printer
The code to take a transaction line id and produce a text representation
of it was present in two places, in slightly different versions: in the
register, for display on the screen, and in the printer module.  The
printer module had a slightly different version that would always use
the 'short' name of an item of stock in the (vain) hope that it wouldn't
run over more than one line on the receipt.

The shared version always uses the full name (manufacturer plus name,
plus ABV in brackets if present, plus unit name).  This often results in
lines that are too long for printed receipts.  To deal with this,
wordwrapping code has been added to the ESC/POS printer driver.  The
food order printer now uses the printer driver's word-wrapper, rather
than doing it itself and making possibly incorrect assumptions about the
receipt width.

TODO: the pdf printer driver should be made to wrap text as well!