Food orders now end up in the transaction log
The "free text" feature added in the previous commit is now used by the
food ordering system to add the full details of food orders to
transactions.  These will be printed on receipts, etc.

The hard-coding of department 10 for food orders has been removed.  The
interface exposed by the register actually allows the food ordering
system to use multiple departments if necessary, but the food ordering
module does not yet support this - menu file format changes would
probably be needed.

The message printed at the bottom of the customer copy of a food order
is now configurable in the menu file.