Combine new/edit/view deliveries options into one menu
There used to be separate menu options for creating a new delivery,
editing an unconfirmed delivery, and viewing an old delivery.  Now there
is just one option, which brings up a list of deliveries both confirmed
and unconfirmed, with an option to create a new delivery as the first
item in the list.  This should help prevent the mistaken duplication of
entry of deliveries.