Quick and easy way to nobble empty transactions
The register now attempts to create a transaction when the "order food"
button is pressed; this is so that an error can be shown (for example
when there is no session active) before the food ordering interface pops
up.

If the user cancels the food ordering interface without placing an
order, there will still be an empty open transaction on the register.
Most of the time this won't be a problem - it will be used the next time
the register is used.  It may be confusing, though, if it is still
hanging around at the end of the session.

Pressing "Cash/Enter" on a null transaction now cancels the transaction.