Proper handling of VAT rate changes (requires database update)
This change makes use of the vatrates table and functions defined in the
previous commit.  VAT receipts are now printed based on information from
the database; the static information in the till configuration file
('vatrate', 'vatno' and 'companyaddr') are no longer used.

The VAT rate and business information for a receipt are chosen based on
the session date of the receipt, so printing of historical receipts
ought to work properly as well.

The example 'bar billiards checker' in the extras file now reads the
current VAT rate for band 'A', which is assumed to be present.