Deal with kitchen printer failures
When access to the kitchen printer fails (usually a network problem, or
the printer not being plugged in or turned on), pop up a helpful message
about what to do.  The food order will already have been printed on the
receipt printer and saved in the current transaction; the customer copy
of the order can be given to the kitchen to make while the printer is
fixed.