Change Cancel key behaviour based on age of line being cancelled
Until now, the behaviour of the Cancel key on an open transaction has
depended only on whether the transaction has been recalled; transactions
that had not been recalled could be deleted in their entirety, or could
have lines removed from them without leaving a log.

Now there is a change in behaviour based on how long ago the lines in
the transaction were created.  Lines that are over one minute old can't
be deleted, and transactions where the first line is greater than one
minute old can't be cancelled all in one go.  Instead, a reversing line
is created and appended to the current transaction.  This may leave the
transaction with a null balance; it can be closed by pressing Enter.