Change the "negative stock remaining" message to be more emphatic
We STILL have problems with staff carrying on selling from a bottle that
the till things is empty - they start the new bottle and don't tell the
till.  For every sale they make a red warning box pops up and they just
ignore it!  Let's change that box so it doesn't just go away when they
press Clear, and is more explicit about them having to get it dealt with
promptly.