Display transaction notes in menus
There is a per-transaction "notes" field (varchar(60)) in the database
which has until now not been used - it was initialised to the name of
the register page that created the transaction, but never read.

This commit changes the default notes to 'null', and adds support for
displaying the notes in lists of transactions on the "recall
transaction" and "merge transactions" functions, and the list of open
transactions that appears at startup.

It is recommended to set all the old transaction notes to 'null' after
installing this version of the software:

UPDATE transactions SET notes=null;