Extra checking on "use stock" for items less than 20% used
One of the most common mistakes made by staff has been for them to mark
a stock item as "finished" when it is actually brand new.  This is
probably because they have picked up a bottle, realised it's still
sealed, and think that the till needs to be told about it when in fact
someone has already done this; they fail to spot that the item they are
marking as "finished" is the one they have in their hand.

This commit adds a check to the "use stock" function: if an item is less
than 20% used, the options to mark it as finished are not made
available.  Instead there is a prompt telling them how to mark it as
finished once it's disconnected from the line, for the rare case that
this is really what they want to do.  If they disconnect the item at
this point, it will still be available in the list of available stock
that pops up next, and they will be able to choose it and reconnect it.