Import twitter and oauth2 libraries
We used to have a "twitter from the till" feature implemented in some
site-specific configuration files - it was a short bit of code making
use of the python-twitter library.

The version of python-twitter in Ubuntu 10.04 LTS (the current target
for the till software) only supports basic authentication, which has now
been disabled by Twitter.  This commit adds a more recent version of the
python-twitter library (accessible from outside the package as
quicktill.twitter) and its dependency python-oauth2.

These libraries will be removed from the till software in the version
that targets the next LTS release of Ubuntu, currently expected to be
version 12.04.