PingaPint support in the register
Add support for vouchers from http://www.pingapint.com/

When pingapint is configured, the "Manage Transaction" menu has a new
option to redeem a voucher.  Vouchers are treated as a new payment type
(alongside cash and cards).

The full response from the pingapint server is stored in a table in the
database for audit purposes.  This table links the response to the
payments table.  A side effect of this is that it is not possible to
cancel a voucher redemption (the PingaPint server doesn't support this).
The user interface does not currently prohibit the attempt to delete
voucher redemptions, but attempts won't succeed.  If this turns out to
be a problem in practice we can add a suitable message.

The right way to do this would be for PingaPint to have an API call to
reverse a validation - make the voucher code valid again.