Fix for input focus ending up on wrong page
It's always been an assumption that a popup box will only be dismissed
through user action.  When it's dismissed, it restores the input focus
to whatever it was at the point the box was popped up.

The alarm popup broke that assumption by dismissing itself after five
minutes.  If the user had switched to a different page, then the input
focus would be restored to the page the alarm popup appeared on.  This
really confused things!

As a temporary fix, the alarm popup refuses to dismiss itself unless it
has the input focus; if it doesn't, it will wait a little while and then
try again.

A more permanent fix would involve management of the input focus being
separated from management of the stack of popups on a page.  It should
be possible to do this with very few changes to the code outside ui.py.