Lock Screen function
This is a popup window that allows a screen to be locked until the user
who locked it re-enters their code.  It is _not_ a security feature; it
is to guard against staff accidentally modifying someone else's
transaction.  It includes a code lookup function on the lock and unlock
pages; if you have a good reason to unlock someone else's transaction
then you can do so.

This feature requires another table in the database; see the "users"
table in the createdb example script.