sqlalchemy ORM description of database
I believe this description of the database is compatible with the
current createdb script; it can be used as a library to read information
from the database.  Updating and inserting has not been tested.
Creating database through the sqlalchemy metadata should yield a
compatible database, although it is missing a couple of views that are
used by the current td.py code.

Additional queries and properties need to be added to the ORM classes to
make this description more useful.

The createdb script is modified in this commit to add a primary key to
stock_annotations, and to update the rule that makes entries in the
stocktype stockline log.  This can be achieved in existing databases
with the following SQL commands:

CREATE SEQUENCE stock_annotation_seq;
ALTER TABLE stock_annotations ADD COLUMN id INTEGER NOT NULL PRIMARY KEY DEFAULT nextval('stock_annotation_seq');
CREATE OR REPLACE RULE log_stocktype AS ON INSERT TO stockonsale
       DO ALSO
       INSERT INTO stockline_stocktype_log VALUES
       (NEW.stocklineid,(SELECT stocktype FROM stock
       	WHERE stock.stockid=NEW.stockid));