Convert data integrity rules into triggers
The various rules we had that end in DO INSTEAD NOTHING broke the
RETURNING keyword.  Note that, because DO INSTEAD NOTHING fails
silently, an attempt to (for example) create a new open session
when one already exists would have resulted in a mapped object
that does not exist in the database.  This would have been
confusing!

This commit rewrites the data integrity rules as triggers which
raise an exception when the rules are violated.  This will be
passed back to sqlalchemy and ultimately the caller, and invalid
mapped objects will not be created.

This commit probably needs more testing!