More sqlalchemy orm model development
Use an abstract base class (or "mixin" in sqlalchemy parlance) for
VatBand and VatRate

Add expression forms for StockItem.used and StockItem.remaining (this
took a lot of work to get right!)  - note that they are not used by
default, you have to name them explicitly as results you expect from a
query(); they don't help when simply accessing used and remaining on an
instance.

Fix up some tillweb URLs

Older versions of postgresql are grumpier about the columns named in
GROUP BY clauses.  Make sure we name all columns in a table when we're
grouping by them.