Convert used and remaining to column properties
Having used and remaining as hybrid properties worked, but meant being
careful with queries: if the expression form was desired it was always
necessary to be explicit.

Turning them into column properties using the column_property() function
means that they will be loaded at the same time as the StockItem object,
AND they are useable in queries.  This is more efficient in the common
case - it's definitely better for the tillweb, and we usually want to
refer to the amount remaining after accessing a StockItem object in the
till.

It may be necessary to add code to invalidate / update StockItem objects
when entries are made or changed in the stockout table.