Convert Session.total to be a deferred column property
Session is accessed a lot in the till software (to determine what the
current session is).  We don't need to work out the session total every
time we do this.

This commit converts the total property from a regular python @property
(which executes a query whenever it's accessed) to a deferred column
property.  The query to load the session total is only executed when
it's accessed, or when it's explicitly undeferred in the session.query
options.