Try to use smaller qrcode in popup window
Linux terminals don't usually have the "upper half block" and "lower
half block" characters that we use to draw larger qrcodes.

Until we can get a set of fonts with these characters in, try using a
smaller qrcode with no label field, which we may be able to draw using
full block characters.