Allow different default font size settings per printer
We had been using font 1 for everything on the TM-U220 to be able to fit
in a sensible number of characters per line.  On the TM-T20 this just
looks squashed.

Use font 0 by default on the TM-T20 instead, and only switch to font 1
for lines that are expected to have a lot of text in.