Merge StockOnSale with StockItem
StockOnSale and StockItem being separate tables is a mistake originally
made in 2004.  This seems like a good time to fix it!

Note that this makes the model defined here incompatible with the
current stock databases (at least until the update script in examples/
is run) and so this model should not be deployed to tillweb.