Pages can now be created dynamically (config change)
Getting ready to implement personal register pages for members of staff
with tokens.  Rearrange the UI code for pages so that they can be
created after startup.

Till-level "hotkeys" now just run functions which are able to do
anything - selecting or creating a new page, popping up a window, etc.