Remove stockref column from translines
The existence of the "stockref" column on translines was a design error
on my part when I wrote the schema in 2004.  What it tried to do can be
accomplished much more easily with a combination of a unique and foreign
key constrant on the "translineid" column in stockout.

When we need to find the stockout row associated with a transaction
line, we just search for the translineid in stockout!  (In the
sqlalchemy model this is a simple backref with uselist=False).

When I extend the schema to support continuous quantities of stock (eg.
for wine and soft drink cartons) the table that records stock movements
can also refer to translines in this way - although getting the database
to enforce a single transline only being referred to by one of these
tables at a time could be tricky!