Add django-based web interface
NB this has been imported from the separate 'ipladmin' project (not yet
open-sourced) and refers to an earlier version of the quicktill.models
module.  It won't work as-is.

The intent here is to add enough django boilerplate that the tillweb
code can be tested and debugged locally, and deployed on single-pub
systems.  For multi-pub systems the tillweb module can be used as a
django app in a larger website.