Merge linepopup into listpopup, remove ui.table()
linepopup and listpopup shared so much code, and linepopup had so few
callers, that it made sense to merge the two and update the callers.

A consequence of this is that the final uses of ui.table() could be
converted to use ui.tableformatter(), so ui.table() has been removed.