Fix up ORM session handling
We can't do the session handling in the handle_keyboard_input method
because it may be called recursively.  Do it in the keyboard driver
instead, and note that other sources of events (eg. a network socket)
will have to manage it too if they call any code that makes use of the
database.