Remove the "annotate_location" function
This was "nag" code intended to remind pub managers to tell the till
where in the cellar real ale was stillaged prior to it being put on
sale.

It turns out, nagging really doesn't help if they aren't routinely using
the stillage management feature of the till.  We ended up with so many
"" locations recorded...

Removing this code closes #5 on github.