Fix focus problem when aborting print of kitchen order
When the register code detects that a transaction may have been edited
on another terminal, it pops up a warning message.  The food order code
was then dismissing the food order popup, returning the focus to the
register but leaving the warning popup in place.

When the warning popup appears, leave the food order popup open.  This
closes #45 on github.