Improve handling of users moving between terminals
When a key is pressed on a register which isn't the user's current one
(according to the entry in models.User.register), the terminal
immediately locks.

Remove the "teleport" function.  The only way for a user to make a
register their current one is now to present their user token.

This closes #48 on github.