Delete register pages based on unsaved data rather than popups
Register pages used to decide whether to delete themselves when
deselected by looking to see if they still had the focus.

It turns out lots of users leave informational popups up ("No Sale",
"Transaction cancelled", and so on), which means their pages don't get
deleted.

Now we allow windows to declare that they are worth keeping around by
adding an attribute "unsaved_data" which is a description of whatever is
unsaved.  The register looks up the focus stack and only permits itself
to be deleted if it finds no unsaved data.