Don't set user's register to None when deleting register page
One of the reasons to delete a register page is that the user is already
working on another terminal.  In this case, the user's register entry in
the database should not be set to None: this would interrupt their work
on the other terminal by locking it on their next keypress.

Even when a user is explicitly deleting their register page, leaving its
guid in their database entry should be harmless: it will be overwritten
next time their user token is seen by any terminal.