Add optional timeout to stock terminal page
If max_unattended_updates is specified then the page will delete itself
(well, invoke the lock screen at the moment; this will have to change
later) after that number of changes without a keypress.

This should probably be passed to the usertoken handler on dedicated
stock terminals.

This closes #46 on github.