Allow whole session list to be viewed
The session list is limited to 100 entries by default for performance
reasons.  (With >4000 sessions, calculating all the totals takes the
database a long time.)

If the list is truncated, add an option at the end to show the whole
list.

Also change the session list code to use ui.tableformatter; the only
reason it wasn't being used before was that the session list code
predated it.

Closes #25 on github.