Implement pending Bitcoin payments
Payment methods can now indicate a payment is 'pending'.  The register
code looks for pending payments when Cash/Enter is pressed and offers
the choice to resume them.  The Bitcoin payment method makes use of
this.

The Bitcoin QR code popup now no longer counts as unsaved data.

This implements #60 on github.