Add "toast" user-interface element
A toast is a message that appears in a popup window and disappears again
with no user intervention.  Its presence does not affect the input
focus; you can carry on pressing keys and they will work as usual.

If multiple toasts are generated then they are shown to the user in turn.

This closes #64 on github.