Ignore usertokens when deciding to lock a register page
This is a speculative fix for #59 on github.  If the hotkey being
processed is a usertoken, skip the check for the user having moved to a
different terminal.  If the usertoken is for someone else then their
page will be deselected anyway.  If it's for them then the page will be
re-selected immediately.

The main change introduced by this fix is that presenting an unknown
user token when a user has moved away from the screen will not
immediately lock the screen; that will happen on the next keypress.

Closes #59 on github - at least for now!