Ensure focus is set to None when page is deselected
Issue #59 on github describes a crash that could arise because the input
focus was not set to None when a page was deselected.  Keypresses could
be passed to the deselected page, which could then crash.

This commit sets the input focus to None when a page is deselected.  If
the page was being deselected because another page is being selected,
that's fine - the new page will set the focus to itself.  If the page is
being deselected because it's dismissing itself, that's also fine -
_ensure_page_exists() will be called at the start of the next time round
the event loop, which will create a page and set the focus.  We just
have to ensure that we don't try to process a keypress when the focus is
None.

Note that this simple fix for issue #59 only works because the popups
that could be thrown up by the usertoken handler are now implemented as
toasts rather than genuine popup dialog boxes.  A toast will happily
remain on screen while the current page is deselected and a new page is
created or selected.

Closes #59 on github, hopefully!