Remove the 'allow_tabs' config option
allow_tabs was always a hack.  It was originally implemented before the
permissions system existed to try to prevent the "Recall Transaction"
button being used to create transactions that were not on display.
Before the usertoken system and linking of transactions with users this
was a problem - a user could return to the till to find the transaction
they were working on "missing", and just create a new one rather than go
back to the one they started.  At the end of the night there would be a
number of transactions with no record of why they were there.

This commit replaces it with a nag screen that insists a user set a
transaction note before they can replace an open transaction on their
display with a different one.

Closes #20 on github.