Flush toasts to the display immediately
Sometimes toasts are displayed just before we do a lengthy or
potentially blocking operation (for example, outputting to the receipt
printer).  Make sure that toasts are flushed to the display immediately;
if we wait until we get back to the event loop the toast's timer might
already have expired and it will never be displayed.