Enable multiple label printers to be configured
Several pubs have more than one label printer - a laser printer
outputting to sheets of A4 with 18 labels per sheet, and a DYMO thermal
printer that prints slowly on individual labels.

Let the user choose where to send the labels; outputting a single label
to an A4 sheet of 18 is silly when there's an alternative!