Cache food order module
The food menu file doesn't change very often.  We had been creating a
new python module and executing the file in it every time; this leaks
memory, because python modules can't be unloaded.  Store a hash of the
menu file, and only create a new module if the file has changed.

Tricks that depended on the module being recreated each time (eg. doing
time-of-day dependent things during initialisation) won't work any more.
If that's what you need, you'll have to add code (eg. a subclass of
foodorder.simplemenu that overrides display_menu) that checks the
time-of-day each time the menu is used.