Add parameter to automenu to specify how options should be spilled
automenu would always produce a scrolling menu if it was passed a list
with more than ten items in it.  This is retained as the default
behaviour, but can be changed by passing a parameter spill="keymenu"
which will make it use a succession of keymenus to manage the excess
items in the list.