Allow default user to be specified on command line
Add a command to list users and IDs.  The IDs may be used with the new
command-line option "-f" or "--user".

If a page (currently only the stock terminal page) needs a user and
hasn't had one supplied, it can call
user.load_user(tillconfig.default_user) to attempt to obtain a user from
the database.

Closes #75 on github.