Don't dismiss the stockline and PLU popups while editing bindings
Editing bindings caused an implicit "dismiss" of the stockline or PLU
popup; this was done without implicitly saving the data and so could
lead to loss of changes entered by the user.

Allow bindings to be added and modified without dismissing the popup;
the editing interface pops up on top, and when done causes a reload and
redisplay of the list of bindings.

Closes #86 on github.