Don't crash when merging with transaction owned by other user
The database is supposed to have a UNIQUE constraint on User.trans_id.

If it does, we used to crash when the current transaction was merged
with a transaction owned by another user, because sqlalchemy would not
set the other user's trans_id to null before setting the current user's
trans_id.

Add a flush to make sure this ordering is explicit.

Some of the till databases don't have this constraint, so have not been
suffering from this crash.  Once this commit is in use, give the
following commands in psql to create the constraint:

CREATE UNIQUE INDEX "users_transid_key" ON users (transid);
ALTER TABLE users ADD UNIQUE USING INDEX "users_transid_key";