Updated stock level check function
Stock level check used to take a single period and use it both for
estimating the rate of stock use and calculating how much stock to buy
based on that estimate.

This updated version allows different periods to be specified for these,
to give a more accurate estimate of stock usage rate.