Split user class has_permission() method into two
Sometimes it's useful to determine whether a user has been given a
permission explicitly, ignoring whether they have it because they are a
superuser.

Split the has_permission() method into two: has_permission() tests for
the permission ignoring the superuser flag, and may() tests for it while
respecting the superuser flag.