Enable exit menu options to be specified on the command line
If no options are specified, defaults to the three that were embedded in
the software prior to this commit.  (Exit codes 0, 2 and 3 for
"restart", "reboot", "turn off").

In a future release, the defaults will be removed and if no options are
specified on the command line the till will exit with code 0 when the
exit menu option is used.