Add option for the till to exit if idle on the lock screen
The till software seems to be a bit memory-leaky, and as far as I can
tell it's not a bug in the python code (the number of uncollected
objects doesn't grow over time).

As a workaround, add an option for the till software to exit if it is
idle and has been running for at least a minimum length of time; the
startup script should spot this and restart.