Change Transline.stockref to be a list
This is the start of the code for the 0.12 branch of quicktill.  It
introduces a change which is incompatible with current databases and
config files.

We are going to support having multiple StockOut objects on a Transline,
to enable (for example) sales of wine which comes from more than one
case.

We remove the UNIQUE constraint on StockOut.translineid and change the
backref from Transline to StockOut to be a list.