Reduce curses dependencies part 1: keycodes
Various parts of the code used raw curses keycodes (integers).  Convert
these either to internal keycode objects or unicode strings in the
keyboard driver.

The number keys were unusual in being internal keycodes that were
printable.  Stop treating these as a special case.  The keymap in the
config file can now map keys to strings directly.

Remove the requirement that menu keys in the keyboard bindings for line
keys have internal keycodes.  If the menu key stored in the database
corresponds to an internal keycode, use that; otherwise use the string
directly.