Allow payment methods to refuse to provide a total
We didn't differentiate between "no total" and "zero total" in the past.
This meant that if a payment method, like Bitcoin, couldn't retrieve a
total because of a network problem, the total would come up as zero instead.

Now that zero means zero, we record it in the database for each payment
method for each session.  This fixes issue #107 on github, where
sessions that genuinely had no total couldn't have their total recorded.