Refactor to move curses-specific code to the ui module
This commit continues the refactor that removes curses-specific code
from the general codebase into the ui module, to make the forthcoming
change to supporting multiple user interface systems easier.

As part of this, it is no longer necessary to specify curseskeyboard()
in the configuration file; the default will always be appropritate for
the user interface system.  Only keyboards that need an extra layer of
processing, like Preh programmable keyboards, need to be specified.